:truck: move auth-method files
