--- src/screens/login/auth-methods/index.ts
@@ -1,3 +1,0 @@
-export { EmailPasswordAuthForm } from "./email-password-form";
-export { StubAuthForm } from "./stub-form";
-export { UnderConstructionForm } from "./under-construction-form";
