+++ src/widgets/auth/forms/email-verification.tsx
