+++ src/widgets/auth/forms/stub-form.tsx
