+++ src/widgets/auth/methods/index.ts
@@ -1,0 +1,3 @@
+export { StubAuthForm as AuthMethodStubWidget } from "../forms/stub-form";
+export { UnderConstructionForm } from "../forms/under-construction-form";
+export { EmailPasswordAuthWidget } from "./email-password";
